Fetching Basic Info using Retrofit in ProfileTab2 fragment
