+++ app/src/main/java/com/appteamnith/hillffair/activities/HomeActivity.java
@@ -52,7 +52,7 @@ public class HomeActivity extends AppCompatActivity
 
         recyclerView = (RecyclerView) findViewById(R.id.list);
 
-        Integer photo[] = {R.drawable.img_1611, R.drawable.img_1610, R.drawable.img_1612,R.drawable.img_1613};
+        Integer photo[] = {R.drawable.wonderwoman, R.drawable.superman, R.drawable.batman,R.drawable.ironman};
         adapter = new HomeAdapter(Arrays.asList(photo),this);
         GridLayoutManager staggeredGridLayoutManager = new GridLayoutManager(this, 3);
         recyclerView.setLayoutManager(staggeredGridLayoutManager);
