+++ app/src/main/java/com/appteamnith/hillffair/utilities/APIINTERFACE.java
@@ -3,7 +3,9 @@ package com.appteamnith.hillffair.utilities;
 import com.appteamnith.hillffair.activities.EventActivity;
 import com.appteamnith.hillffair.activities.ProfileActivity;
 import com.appteamnith.hillffair.activities.UploadNewsFeedActivity;
+import com.appteamnith.hillffair.fragments.ProfileTab2;
 import com.appteamnith.hillffair.modals.Login;
+import com.appteamnith.hillffair.modals.ProfileDataModel;
 import com.appteamnith.hillffair.modals.Register;
 import com.appteamnith.hillffair.modals.newsfeed_model;
 
@@ -39,7 +41,7 @@ public interface APIINTERFACE {
 
     @FormUrlEncoded
     @POST("profile")
-    Call<ProfileActivity.ProfileBasicDetailModel> profileBasicInfo(@Field("id") String id);
+    Call<ProfileDataModel> profileBasicInfo(@Field("id") String id);
 
 
     @GET("club")
