+++ app/src/main/res/drawable/batman.JPG
