+++ app/src/main/res/layout/activity_theme_selection.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:orientation="vertical"
+    android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_marginLeft="10dp"
     android:layout_marginRight="10dp"
@@ -44,6 +46,7 @@
                         android:src="@drawable/comic9" />
 
                     <ImageButton
+
                         android:id="@+id/flash"
                         android:layout_width="match_parent"
                         android:layout_height="110dp"
@@ -60,7 +63,8 @@
                     android:orientation="horizontal">
 
 
-                    <ImageButton
+                    <ImageButton xmlns:tools="http://schemas.android.com/tools"
+                        tools:ignore="ContentDescription"
                         android:id="@+id/superman"
                         android:layout_width="match_parent"
                         android:layout_height="110dp"
@@ -68,7 +72,8 @@
                         android:layout_weight="0.5"
                         android:src="@drawable/comic2" />
 
-                    <ImageButton
+                    <ImageButton xmlns:tools="http://schemas.android.com/tools"
+                        tools:ignore="ContentDescription"
                         android:id="@+id/captain"
                         android:layout_width="match_parent"
                         android:layout_height="110dp"
