+++ app/src/main/res/layout/profile_tab2_fragment.xml
@@ -6,6 +6,8 @@
     <ScrollView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:visibility="gone"
+        android:id="@+id/scroll_layout"
         >
         <LinearLayout
             android:layout_width="fill_parent"
@@ -63,26 +65,6 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content">
                 <EditText
-                    android:id="@+id/et3"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:singleLine="true"
-                    android:layout_margin="5dp"
-                    android:background="#fff"
-                    android:clickable="true"
-                    android:cursorVisible="false"
-                    android:focusable="false"
-                    android:focusableInTouchMode="false"
-                    android:elevation="4dp"
-                    android:padding="10dp"
-                    android:text="B.Tech CSE"
-                    android:hint="Branch"/>
-            </android.support.design.widget.TextInputLayout>
-
-            <android.support.design.widget.TextInputLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-                <EditText
                     android:id="@+id/et4"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -126,16 +108,12 @@
 
     </ScrollView>
 
-    <ImageButton
-        android:onClick="btnClick"
-        android:id="@+id/edit_btn"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentEnd="true"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="10dp"
-        android:background="@drawable/profile_edit_btn_bg"
-        android:src="@android:drawable/ic_menu_edit"/>
+   <ProgressBar
+       android:layout_width="wrap_content"
+       android:layout_height="wrap_content"
+       android:layout_centerInParent="true"
+       android:visibility="gone"
+       android:id="@+id/progress"/>
